requirements.txt
fpdf
python-docx
reportlab
google-generativeai
Pillow